chore(jest): update jest tests
