--- ponicode.config.json
@@ -1,11 +1,0 @@
-{
-    "schemaVersion": "1",
-    "javascript": {
-        "name": "Core-PWA",
-        "rootDir": "./",
-        "testEnvironment": "node",
-        "testLocationStrategy": "relativeToProjectRoot",
-        "testLocationFolder": "./ponicode",
-        "testNamePattern": "{name}.test.{ext}"
-    }
-}
