--- ponicode/store/iridium/state.test.ts
@@ -1,8 +1,0 @@
-import * as state from '~/store/iridium/state'
-// @ponicode
-describe('state.default', () => {
-  test('0', () => {
-    const result: any = state.default()
-    expect(result).toMatchSnapshot()
-  })
-})
